refactor download_json into utils.json_utils
