+++ datagovharvester/extract/__init__.py
@@ -1,6 +1,6 @@
 from datagovharvester.extract.dcatus import parse_catalog
-from datagovharvester.extract.utils import download_json
 from datagovharvester.utils.s3_utilities import upload_to_S3
+from datagovharvester.utils.json_utilities import download_json
 from datagovharvester.validate.dcat_us import is_dcatus_schema
 
 # ruff: noqa: F841
